TalentOS/requirements.txt
# Django core
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0

# Auth
djangorestframework-simplejwt>=5.3,<6.0
PyJWT>=2.8,<3.0
pyotp>=2.9,<3.0
qrcode[pil]>=7.4,<8.0

# Database
psycopg2-binary>=2.9,<3.0

# Caching
django-redis>=5.4,<6.0

# Celery
celery[redis]>=5.4,<6.0
django-celery-beat>=2.6,<3.0

# Storage
boto3>=1.34,<2.0
django-storages>=1.14,<2.0

# Search
opensearch-py>=2.5,<3.0

# Resume parsing
pdfplumber>=0.11,<1.0
python-docx>=1.1,<2.0
pytesseract>=0.3,<1.0
Pillow>=10.0,<11.0

# NLP (local only)
spacy>=3.7,<4.0

# Text processing
python-dateutil>=2.9,<3.0
dateparser>=1.2,<2.0
jellyfish>=1.0,<2.0

# Utilities
python-decouple>=3.8,<4.0
uuid7>=0.1,<1.0
gunicorn>=22.0,<23.0

# Monitoring / Observability
sentry-sdk[django]>=1.40,<3.0
django-prometheus>=2.3,<3.0

# Markup
django-simple-history>=3.5,<4.0
Markdown>=3.6,<4.0

# Export
openpyxl>=3.1,<4.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
factory-boy>=3.3,<4.0
pytest-cov>=5.0,<7.0

stripe>=11.0,<15.0
drf-spectacular>=0.27,<1.0
cryptography>=42.0,<44.0
bleach>=6.1,<7.0

# DNS / domain verification
dnspython>=2.6,<3.0

# Language detection & translation
langdetect>=1.0,<2.0
deep-translator>=1.11,<2.0

# SMS / voice messaging
twilio>=8.0,<10.0

# File type detection
python-magic>=0.4,<1.0

# ConnectOS shared platform layer
connectos-platform @ file:../../ConnectOS/platform
